Update "creating an adapter" tutorial
