+++ docs/tutorials/creating-an-adapter/05-Historical_Value_Queries.md
@@ -131,7 +131,7 @@ The `ReadHistoricalTagValues` class implements all functions defined in `Default
 
 ## Testing
 
-Modify the `Run` method in `Program.cs` as follows:
+Modify the `Run` method in `Runner.cs` as follows:
 
 ```csharp
 private static async Task Run(IAdapterCallContext context, CancellationToken cancellationToken) {
