+++ examples/tutorials/Directory.Build.props
@@ -1,5 +1,6 @@
 ﻿<Project>
   <PropertyGroup>
-    <AdapterPackageVersion>1.0.2</AdapterPackageVersion>
+    <AdapterPackageVersion>2.0.0-beta.408</AdapterPackageVersion>
+    <MicrosoftExtensionsHostingVersion>5.0.0</MicrosoftExtensionsHostingVersion>
   </PropertyGroup>
 </Project>
