+++ examples/tutorials/Directory.Build.targets
@@ -1,6 +1,6 @@
 <Project>
   <ItemGroup Condition=" '$(AdapterPackageVersion)' == '' ">
     <PackageReference Remove="IntelligentPlant.AppStoreConnect.Adapter" />
-    <ProjectReference Include="..\..\..\src\DataCore.Adapter\DataCore.Adapter.csproj" /> 
+    <ProjectReference Include="$(MSBuildThisFileDirectory)..\..\src\DataCore.Adapter\DataCore.Adapter.csproj" /> 
   </ItemGroup>
 </Project>
