+++ examples/tutorials/creating-an-adapter/chapter-01/Adapter.cs
@@ -16,7 +16,7 @@ namespace MyAdapter {
             string description = null,
             IBackgroundTaskService backgroundTaskService = null,
             ILogger<Adapter> logger = null
-        ) : base(id, name, description, backgroundTaskService, logger) { }
+        ) : base(id, new AdapterOptions() { Name = name, Description = description }, backgroundTaskService, logger) { }
 
 
         protected override Task StartAsync(CancellationToken cancellationToken) {
