+++ examples/tutorials/creating-an-adapter/chapter-02/Adapter.cs
@@ -18,7 +18,7 @@ namespace MyAdapter {
             string description = null,
             IBackgroundTaskService backgroundTaskService = null,
             ILogger<Adapter> logger = null
-        ) : base(id, name, description, backgroundTaskService, logger) { }
+        ) : base(id, new AdapterOptions() { Name = name, Description = description }, backgroundTaskService, logger) { }
 
 
         private static DateTime CalculateSampleTime(DateTime queryTime) {
