+++ examples/tutorials/creating-an-adapter/chapter-02/Program.cs
@@ -1,70 +1,22 @@
-﻿using System;
-using System.Threading;
-using System.Threading.Tasks;
-using DataCore.Adapter;
-using DataCore.Adapter.Diagnostics;
-using DataCore.Adapter.RealTimeData;
+﻿using System.Threading.Tasks;
+
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace MyAdapter {
     class Program {
 
-        private const string AdapterId = "example";
-
-        private const string AdapterDisplayName = "Example Adapter";
-
-        private const string AdapterDescription = "Example adapter, built using the tutorial on GitHub";
-
-
         public static async Task Main(params string[] args) {
-            using (var userCancelled = new CancellationTokenSource()) {
-                Console.CancelKeyPress += (sender, e) => userCancelled.Cancel();
-                try {
-                    Console.WriteLine("Press CTRL+C to quit");
-                    await Run(new DefaultAdapterCallContext(), userCancelled.Token);
-                }
-                catch (OperationCanceledException) { }
-            }
+            await CreateHostBuilder(args).RunConsoleAsync().ConfigureAwait(false);
         }
 
 
-        private static async Task Run(IAdapterCallContext context, CancellationToken cancellationToken) {
-            await using (IAdapter adapter = new Adapter(AdapterId, AdapterDisplayName, AdapterDescription)) {
-
-                await adapter.StartAsync(cancellationToken);
-
-                Console.WriteLine();
-                Console.WriteLine($"[{adapter.Descriptor.Id}]");
-                Console.WriteLine($"  Name: {adapter.Descriptor.Name}");
-                Console.WriteLine($"  Description: {adapter.Descriptor.Description}");
-                Console.WriteLine("  Properties:");
-                foreach (var prop in adapter.Properties) {
-                    Console.WriteLine($"    - {prop.Name} = {prop.Value}");
-                }
-                Console.WriteLine("  Features:");
-                foreach (var feature in adapter.Features.Keys) {
-                    Console.WriteLine($"    - {feature}");
-                }
-
-                var readSnapshotFeature = adapter.GetFeature<IReadSnapshotTagValues>();
-                var snapshotValues = await readSnapshotFeature.ReadSnapshotTagValues(
-                    context,
-                    new ReadSnapshotTagValuesRequest() { 
-                        Tags = new[] { 
-                            "Example 1",
-                            "Example 2"
-                        }
-                    },
-                    cancellationToken
-                );
-
-                Console.WriteLine();
-                Console.WriteLine("  Snapshot Values:");
-                await foreach (var value in snapshotValues.ReadAllAsync(cancellationToken)) {
-                    Console.WriteLine($"    [{value.TagName}] - {value.Value}");
-                }
-
-            }
+        private static IHostBuilder CreateHostBuilder(string[] args) {
+            return Host.CreateDefaultBuilder(args).ConfigureServices(services => {
+                services.AddHostedService<Runner>();
+            });
         }
 
     }
+
 }
