+++ examples/tutorials/creating-an-adapter/chapter-05/Adapter.cs
@@ -27,16 +27,15 @@ namespace MyAdapter {
             IBackgroundTaskService backgroundTaskService = null,
             ILogger<Adapter> logger = null
         ) : base(
-            id, 
-            name, 
-            description, 
+            id,
+            new AdapterOptions() { Name = name, Description = description },
             backgroundTaskService, 
             logger
         ) {
-            AddFeature<ISnapshotTagValuePush, PollingSnapshotTagValuePush>(PollingSnapshotTagValuePush.ForAdapter(
-                this, 
-                TimeSpan.FromSeconds(1)
-            ));
+            AddFeatures(new PollingSnapshotTagValuePush(this, new PollingSnapshotTagValuePushOptions() {
+                PollingInterval = TimeSpan.FromSeconds(1),
+                TagResolver = SnapshotTagValuePush.CreateTagResolverFromAdapter(this)
+            }, BackgroundTaskService, Logger));
 
             AddFeatures(ReadHistoricalTagValues.ForAdapter(this));
         }
@@ -53,20 +52,13 @@ namespace MyAdapter {
                 ++i;
                 var tagId = i.ToString();
                 var tagName = string.Concat(waveType, "_Wave");
-                var tagProperties = new[] {
-                    CreateWaveTypeProperty(waveType)
-                };
 
-                var tag = new TagDefinition(
-                    tagId,
-                    tagName,
-                    $"A tag that returns a {waveType.ToLower()} wave value",
-                    null,
-                    VariantType.Double,
-                    null,
-                    tagProperties,
-                    null
-                );
+                var tag = TagDefinitionBuilder
+                    .Create(tagId, tagName)
+                    .WithDescription($"A tag that returns a {waveType.ToLower()} wave value")
+                    .WithDataType(VariantType.Double)
+                    .WithProperties(CreateWaveTypeProperty(waveType))
+                    .Build();
 
                 _tagsById[tag.Id] = tag;
                 _tagsByName[tag.Name] = tag;
